refactor: Update nginx configuration to proxy API requests
This commit updates the nginx configuration file to include a new location block for proxying API requests to the backend server. The previous configuration only had a server block for serving static files. With this change, any requests to the /api path will be proxied to the backend server running at http://api-metas:4020/api/v1. This will enable the frontend to communicate with the backend API.